requirements.txt
h5py==3.7.0
libmr==0.1.9
matplotlib==3.5.1
numpy==1.22.3
pandas==1.4.2
Pillow==9.3.0
scikit_learn==1.1.3
scipy==1.8.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0
